backend/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
bitarray==2.7.3
celery==5.2.7
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cytoolz==0.12.1
Django==4.2
django-celery-results==2.5.0
django-cors-headers==3.14.0
django-rest-passwordreset==1.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
ecdsa==0.18.0
eth-abi==4.0.0
eth-account==0.8.0
eth-hash==0.5.1
eth-keyfile==0.6.1
eth-keys==0.4.0
eth-rlp==0.3.0
eth-typing==3.3.0
eth-utils==2.1.0
frozenlist==1.3.3
hexbytes==0.3.0
idna==3.4
importlib-resources==5.12.0
jsonschema==4.17.3
kombu==5.2.4
lru-dict==1.1.8
multidict==6.0.4
parsimonious==0.9.0
pkgutil-resolve-name==1.3.10
prompt-toolkit==3.0.38
protobuf==4.22.3
psycopg2-binary==2.9.6
pycryptodome==3.17
PyJWT==2.6.0
pyrsistent==0.19.3
pytz==2023.3
redis==4.5.4
regex==2023.3.23
requests==2.28.2
rlp==3.0.0
six==1.16.0
sqlparse==0.4.4
toolz==0.12.0
urllib3==1.26.15
vine==5.0.0
wcwidth==0.2.6
web3==6.2.0
websockets==11.0.2
yarl==1.8.2
zipp==3.15.0
